# glad-main/doc/swagger/AoL-Teacher.yaml
openapi: 3.0.0
info:
  title: AoL Teacher Interface
  description: |-
    This is an AoL Teacher API specification.

  contact:
    email: sudhagar@abovecloud9.ai
  version: 0.0.3
...
servers:
  - url: https://api.abovecloud9.ai/api/v1 # placeholder
...
paths:
  /course:
    put:
      tags:
        - course
      summary: Updates an existing course
      description: Update an existing course by course id
      operationId: updateCourse
      requestBody:
        description: Update an existent course
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Invalid ID supplied
        '404':
          description: Course not found
        '405':
          description: Validation exception
      security:
        - bearer: []
    post:
      tags:
        - course
      summary: Adds a new course
      description: Add a new course
      operationId: addCourse
      requestBody:
        description: Create a new course
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '405':
          description: Invalid input
      security:
        - bearer: []
...
  /course/{courseId}:
    get:
      tags:
        - course
      summary: Finds course by ID
      description: Returns a single course
      operationId: getCourseById
      parameters:
        - name: courseId
          in: path
          description: ID of course to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Invalid ID supplied
        '404':
          description: Course not found
      security:
        - bearer: []
    delete:
      tags:
        - course
      summary: Deletes a course
      description: delete a course
      operationId: deleteCourse
      parameters:
        - name: courseId
          in: path
          description: Course id to delete
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid Course id
      security:
        - bearer: []
...
components:
  schemas:
...
    DateTime:
      type: object
      properties:
        date:
          type: string
          format: date
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
    Timezone:
      type: string
      description: Supported timezones
      example: EST
      enum:
        - EST
        - CST
        - MST
        - PST
    Course:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10000000
        extId:
          type: string
          example: 'co0198772'
        name:
          type: string
        organizers:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        contacts:
          description: Contact person for the course
          type: array
          items:
            $ref: '#/components/schemas/Account'  
        teachers:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        notes:
          type: string
          description: Course notes
        status:
          type: string
          description: Course status
          example: open
          enum:
            - open
            - expense-submitted
            - expense-declined
            - closed
            - active
            - declined
            - submitted
            - canceled
            - inactive
        maxAttendees:
          type: integer
          format: int32
          example: 10
        dates:
          type: array
          items:
            $ref: '#/components/schemas/DateTime'
        timezone:
          $ref: '#/components/schemas/Timezone'
        location:
          $ref: '#/components/schemas/Address'
        center:
          $ref: '#/components/schemas/Center'
        type:
          type: string
          description: Course delivery type
          example: in-person
          enum:
            - in-person
            - online
        notify:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        numAttendees: # Cannot be more than maxAttendees
          type: integer
          format: int32
          example: 10
        attendees:
          type: array
          items:
            $ref: '#/components/schemas/User'
        autoApprove:
          type: boolean
...
    Center: # TBD: Should we have CenterDetail and Center?
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 20000000
        extId:
          type: string
          example: 'ce0198772'
        name:
          type: string
          example: 'Happiness Center'
        location:
          $ref: '#/components/schemas/Address'
        status:
          type: string
          description: Center Status
          example: active
          enum:
            - active
            - inactive
        type:
          type: string
          description: Center Type
          example: ashram
          enum:
            - ashram
            - temple-of-knowledge
...
    Address:
      type: object
      properties:
        street:
          type: string
          example: 437 Lytton
        city:
          type: string
          example: Palo Alto
        state:
          type: string
          example: CA
        zip:
          type: string
          example: '94301'
        country:
          type: string
          example: USA
...
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 30000000
        extId:
          type: string
          example: 'ac1977'
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        type:
          type: string
          description: User Type
          example: teacher
          enum:
            - teacher
            - assistant-teacher
            - organizer
            - member
            - user
    User:
      type: object
      properties:
        account:
          $ref: '#/components/schemas/Account'
        username:
          type: string
          example: theUser
        email:
          type: string
          example: john@email.com
        password: # TBD - Likely to be removed
          type: string
          example: '12345'
        phone:
          type: string
          example: '********'
...
  securitySchemes:
    bearer:
      type: apiKey
      name: Bearer
      in: header
    sample_auth_unused:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://abovecloud9.ai/oauth/authorize
          scopes:
            write:courses: modify courses
            read:courses: read courses